Extract resolveGraphQLTypeName to shared lib for renderer reuse
Moves the scalar→GraphQL name mapping (string→String, int32→Int, etc.)
and the type.name fallback logic into src/lib/graphql-type-name.ts.
printMutatedType now imports from there. Renderer components will use
the same function for <gql.Field type={...}>.
